extract tx queue helpers
