+++ loraham_kiss_tnc/bridge.c
@@ -1,6 +1,7 @@
 #include "bridge.h"
 
 #include "bridge_conf.h"
+#include "bridge_tx_queue.h"
 #include "ax25.h"
 #include "kiss.h"
 #include "loraham_sock.h"
@@ -28,7 +29,6 @@
 #define LHKT_RECONNECT_DELAY_SEC 1
 #define LHKT_CLIENT_WRITE_TIMEOUT_SEC 2
 #define LHKT_TX_RESTORE_RETRY_DELAY_MS 100
-#define LHKT_TX_QUEUE_MAX 16
 #define LHKT_TX_START_WAIT_MS 500
 #define LHKT_QUEUE_POLL_USEC 100000
 #define LHKT_TEST_HOOK_MAX_CALLS 16
@@ -649,104 +649,6 @@ static void print_stats_if_due(const lhkt_config_t *cfg,
     *next_stats = now + cfg->stats_interval;
 }
 
-/* ---- TX queue ---- */
-
-#define BRIDGE_TX_DECISION_WAIT  0
-#define BRIDGE_TX_DECISION_SEND  1
-#define BRIDGE_TX_DECISION_DROP  2
-
-typedef struct {
-    uint8_t packet[LHKT_LORAHAM_TX_MAX];
-    size_t packet_len;
-    long queued_ms;
-    long tx_wait_start_ms;
-    long cad_wait_start_ms;
-    long cad_idle_since_ms;
-    int cad_was_busy;
-} bridge_tx_item_t;
-
-typedef struct {
-    bridge_tx_item_t items[LHKT_TX_QUEUE_MAX];
-    size_t head;
-    size_t count;
-} bridge_tx_queue_t;
-
-static void bridge_tx_queue_init(bridge_tx_queue_t *queue)
-{
-    if (!queue) {
-        return;
-    }
-
-    memset(queue, 0, sizeof(*queue));
-}
-
-static size_t bridge_tx_queue_limit(const lhkt_config_t *cfg)
-{
-    if (!cfg || cfg->tx_queue_len < 1) {
-        return 1;
-    }
-
-    if (cfg->tx_queue_len > LHKT_TX_QUEUE_MAX) {
-        return LHKT_TX_QUEUE_MAX;
-    }
-
-    return (size_t)cfg->tx_queue_len;
-}
-
-static int bridge_tx_queue_empty(const bridge_tx_queue_t *queue)
-{
-    return !queue || queue->count == 0;
-}
-
-static bridge_tx_item_t *bridge_tx_queue_head(bridge_tx_queue_t *queue)
-{
-    if (!queue || queue->count == 0) {
-        return NULL;
-    }
-
-    return &queue->items[queue->head];
-}
-
-static void bridge_tx_queue_pop(bridge_tx_queue_t *queue)
-{
-    if (!queue || queue->count == 0) {
-        return;
-    }
-
-    memset(&queue->items[queue->head], 0, sizeof(queue->items[queue->head]));
-    queue->head = (queue->head + 1) % LHKT_TX_QUEUE_MAX;
-    queue->count--;
-}
-
-static int bridge_tx_queue_push(bridge_tx_queue_t *queue,
-                                const lhkt_config_t *cfg,
-                                const uint8_t *packet,
-                                size_t packet_len)
-{
-    bridge_tx_item_t *item;
-    size_t index;
-
-    if (!queue || !cfg || !packet || packet_len == 0 ||
-        packet_len > LHKT_LORAHAM_TX_MAX) {
-        return LHKT_ERR_FORMAT;
-    }
-
-    if (queue->count >= bridge_tx_queue_limit(cfg)) {
-        return LHKT_ERR_NOSPACE;
-    }
-
-    index = (queue->head + queue->count) % LHKT_TX_QUEUE_MAX;
-    item = &queue->items[index];
-    memset(item, 0, sizeof(*item));
-
-    memcpy(item->packet, packet, packet_len);
-    item->packet_len = packet_len;
-    item->queued_ms = bridge_now_ms();
-    queue->count++;
-
-    return LHKT_OK;
-}
-
 static int bridge_wait_tx_complete(int conf_fd,
                                    bridge_conf_state_t *conf_state,
                                    const lhkt_config_t *cfg)
@@ -856,114 +758,6 @@ int lhkt_test_bridge_wait_tx_complete_events(const char *events)
     close(sv[0]);
     close(sv[1]);
     return ret;
-}
-#endif
-
-static int bridge_tx_head_decision(const lhkt_config_t *cfg,
-                                   bridge_conf_state_t *conf_state,
-                                   bridge_tx_item_t *item,
-                                   long now)
-{
-    if (!cfg || !item) {
-        return BRIDGE_TX_DECISION_DROP;
-    }
-
-    if (cfg->tx_packet_ttl_ms > 0 &&
-        now - item->queued_ms >= cfg->tx_packet_ttl_ms) {
-        printf("[TXQ] Drop expired packet len=%zu\n", item->packet_len);
-        return BRIDGE_TX_DECISION_DROP;
-    }
-
-    if (conf_state && conf_state->tx_busy) {
-        if (item->tx_wait_start_ms == 0) {
-            item->tx_wait_start_ms = now;
-        }
-
-        if (now - item->tx_wait_start_ms >= cfg->tx_busy_timeout_ms) {
-            printf("[TXQ] Drop packet after TX busy timeout len=%zu\n",
-                   item->packet_len);
-            return BRIDGE_TX_DECISION_DROP;
-        }
-
-        return BRIDGE_TX_DECISION_WAIT;
-    }
-
-    item->tx_wait_start_ms = 0;
-
-    if (!cfg->cad_ignore && conf_state) {
-        if (conf_state->cad_busy) {
-            item->cad_was_busy = 1;
-            item->cad_idle_since_ms = 0;
-
-            if (item->cad_wait_start_ms == 0) {
-                item->cad_wait_start_ms = now;
-            }
-
-            if (now - item->cad_wait_start_ms < cfg->cad_wait_ms) {
-                return BRIDGE_TX_DECISION_WAIT;
-            }
-
-            printf("[TXQ] CAD busy timeout, sending anyway len=%zu\n",
-                   item->packet_len);
-            return BRIDGE_TX_DECISION_SEND;
-        }
-
-        if (item->cad_was_busy) {
-            if (item->cad_idle_since_ms == 0) {
-                item->cad_idle_since_ms = now;
-                return BRIDGE_TX_DECISION_WAIT;
-            }
-
-            if (now - item->cad_idle_since_ms < cfg->cad_idle_ms) {
-                return BRIDGE_TX_DECISION_WAIT;
-            }
-        }
-    }
-
-    return BRIDGE_TX_DECISION_SEND;
-}
-
-#ifdef LHKT_TEST
-int lhkt_test_bridge_tx_decision(int tx_busy,
-                                  int cad_busy,
-                                  int cad_ignore,
-                                  long queued_age_ms,
-                                  long tx_wait_age_ms,
-                                  long cad_wait_age_ms,
-                                  long cad_idle_age_ms,
-                                  int cad_was_busy)
-{
-    lhkt_config_t cfg;
-    bridge_conf_state_t conf_state;
-    bridge_tx_item_t item;
-    long now;
-
-    lhkt_config_defaults(&cfg);
-    cfg.cad_ignore = cad_ignore;
-    bridge_conf_state_init(&conf_state);
-    memset(&item, 0, sizeof(item));
-
-    conf_state.tx_busy = tx_busy;
-    conf_state.cad_busy = cad_busy;
-    item.packet_len = 10;
-    item.cad_was_busy = cad_was_busy;
-
-    now = bridge_now_ms();
-    item.queued_ms = now - (queued_age_ms >= 0 ? queued_age_ms : 0);
-
-    if (tx_wait_age_ms >= 0) {
-        item.tx_wait_start_ms = now - tx_wait_age_ms;
-    }
-
-    if (cad_wait_age_ms >= 0) {
-        item.cad_wait_start_ms = now - cad_wait_age_ms;
-    }
-
-    if (cad_idle_age_ms >= 0) {
-        item.cad_idle_since_ms = now - cad_idle_age_ms;
-    }
-
-    return bridge_tx_head_decision(&cfg, &conf_state, &item, now);
 }
 #endif
