+++ loraham_kiss_tnc/build.sh
@@ -57,6 +57,7 @@ sources=(
   main.c
   bridge.c
   bridge_conf.c
+  bridge_tx_queue.c
   cli.c
   tcp_server.c
   kiss.c
