+++ loraham_kiss_tnc/run_tests.sh
@@ -154,6 +154,7 @@ run_test bridge \
   "$script_dir/loraham_sock.c" \
   "$script_dir/bridge.c" \
   "$script_dir/bridge_conf.c" \
+  "$script_dir/bridge_tx_queue.c" \
   "$script_dir/ax25.c" \
   "$script_dir/tnc2.c" \
   "$script_dir/kiss.c" \
